docs(docsite): dedupe idPrefix demo via shared re-exported story
- Extract canonical ColorIdPrefix demo into its own module
- Re-export it from both Color Variants and Recipes (single source, 1:1)
- Remove divergent Recipes ScopingGradientIds story
